Refactored activity service
