+++ module/Activity/src/Activity/Service/ActivityQuery.php
@@ -1,0 +1,195 @@
+<?php
+
+namespace Activity\Service;
+
+use Application\Service\AbstractAclService;
+use Activity\Model\Activity as ActivityModel;
+use Zend\ServiceManager\ServiceManagerAwareInterface;
+
+class ActivityQuery extends AbstractAclService implements ServiceManagerAwareInterface 
+{
+    /**
+     * Get the ACL.
+     *
+     * @return \Zend\Permissions\Acl\Acl
+     */
+    public function getAcl()
+    {
+        return $this->getServiceManager()->get('activity_acl');
+    }
+
+    /**
+     * Get the default resource ID.
+     *
+     * This is used by {@link isAllowed()} when no resource is specified.
+     *
+     * @return string
+     */
+    protected function getDefaultResourceId()
+    {
+        return 'activity';
+    }
+    
+
+    /**
+     * Get the information of one activity from the database.
+     *
+     * @param int $id The activity id to be searched for
+     *
+     * @return \Activity\Model\Activity Activity or null if the activity does not exist
+     */
+    public function getActivity($id)
+    {
+        if (!$this->isAllowed('view', 'activity')) {
+            $translator = $this->getTranslator();
+            throw new \User\Permissions\NotAllowedException(
+                $translator->translate('You are not allowed to view the activities')
+            );
+        }
+
+        $activityMapper = $this->getServiceManager()->get('activity_mapper_activity');
+        $activity = $activityMapper->getActivityById($id);
+
+        return $activity;
+    }
+
+    /**
+     * Get the activity with additional details
+     *
+     * @param $id
+     * @return ActivityModel
+     */
+    public function getActivityWithDetails($id)
+    {
+        if (! $this->isAllowed('viewDetails', 'activity')) {
+            $translator = $this->getTranslator();
+            throw new \User\Permissions\NotAllowedException(
+                $translator->translate('You are not allowed to view the activities')
+            );
+        }
+        return $this->getActivity($id   );
+    }
+
+    /**
+     * Returns an array of all activities.
+     * NB: This method is currently unused. Should it be removed?
+     * @return array Array of activities
+     */
+    public function getAllActivities()
+    {
+        if (!$this->isAllowed('view', 'activity')) {
+            $translator = $this->getTranslator();
+            throw new \User\Permissions\NotAllowedException(
+                $translator->translate('You are not allowed to view the activities')
+            );
+        }
+
+        $activityMapper = $this->getServiceManager()->get('activity_mapper_activity');
+        $activity = $activityMapper->getAllActivities();
+
+        return $activity;
+    }
+
+    /**
+     * Get all the activities that are yet to be approved
+     *
+     * @return array Array of activities
+     */
+    public function getUnapprovedActivities()
+    {
+        if (!$this->isAllowed('viewUnapproved', 'activity')) {
+            $translator = $this->getTranslator();
+            throw new \User\Permissions\NotAllowedException(
+                $translator->translate('You are not allowed to view the activities')
+            );
+        }
+
+        $activityMapper = $this->getServiceManager()->get('activity_mapper_activity');
+        $activity = $activityMapper->getUnapprovedActivities();
+
+        return $activity;
+    }
+
+    /**
+     * Get all activities that are approved by the board
+     *
+     * @return array Array of activities
+     */
+    public function getApprovedActivities()
+    {
+        if (!$this->isAllowed('view', 'activity')) {
+            $translator = $this->getTranslator();
+            throw new \User\Permissions\NotAllowedException(
+                $translator->translate('You are not allowed to view approved the activities')
+            );
+        }
+
+        $activityMapper = $this->getServiceManager()->get('activity_mapper_activity');
+        $activity = $activityMapper->getApprovedActivities();
+
+        return $activity;
+    }
+
+    /**
+     * Get upcoming activities organized by the given organ.
+     *
+     * @param \Organ\Model\Organ $organ
+     * @param integer $count
+     *
+     * @return array
+     */
+    public function getOrganActivities($organ, $count = null)
+    {
+        return $this->getActivityMapper()->getUpcomingActivities($count, $organ);
+    }
+
+    /**
+     * Get all activities that are disapproved by the board
+     *
+     * @return array Array of activities
+     */
+    public function getDisapprovedActivities()
+    {
+        if (!$this->isAllowed('viewDisapproved', 'activity')) {
+            $translator = $this->getTranslator();
+            throw new \User\Permissions\NotAllowedException(
+                $translator->translate('You are not allowed to view the disapproved activities')
+            );
+        }
+
+        $activityMapper = $this->getServiceManager()->get('activity_mapper_activity');
+        $activity = $activityMapper->getDisapprovedActivities();
+
+        return $activity;
+    }
+
+    /**
+     * Get all activities that are approved by the board and which occur in the future
+     *
+     * @return array Array of activities
+     */
+    public function getUpcomingActivities()
+    {
+        if (!$this->isAllowed('view', 'activity')) {
+            $translator = $this->getTranslator();
+            throw new \User\Permissions\NotAllowedException(
+                $translator->translate('You are not allowed to view upcoming the activities')
+            );
+        }
+
+        $activityMapper = $this->getServiceManager()->get('activity_mapper_activity');
+        $activity = $activityMapper->getUpcomingActivities();
+
+        return $activity;
+    }
+    
+    /**
+     * Get the activity mapper.
+     *
+     * @return \Activity\Mapper\Activity
+     */
+    public function getActivityMapper()
+    {
+        return $this->sm->get('activity_mapper_activity');
+    }
+}
